TypeX: replace TrueType and FalseType with ConstantBooleanType
